Split the code into several more components
