+++ src/components/ApplicationWindow/Window.tsx
@@ -1,13 +1,12 @@
 import React, { FC } from 'react';
 import styled from 'styled-components';
-import TitleBar from './TitleBar';
+import WindowBar from './WindowBar';
 
 const Container = styled.div`
   background: #c3c3c3;
 
   border-top: 4px solid #eee;
   border-left: 4px solid #eee;
-
   border-right: 2px solid #c3c3c3;
   border-bottom: 2px solid #c3c3c3;
 
@@ -15,22 +14,18 @@ const Container = styled.div`
 
   display: grid;
   grid-template-columns: repeat(12, 8.25%);
-
   padding: 0.5em;
-
-  /* position: absolute; */
   transition: all 2s;
-
   height: 100%;
 `;
 
 type WindowProps = {
   className?: string;
   name: string;
-  onClick: Function;
+  onClick: () => void;
   onClose?: () => void;
   // eslint-disable-next-line no-undef
-  children: JSX.Element[] | JSX.Element;
+  children: JSX.Element | JSX.Element[];
 }
 
 const Window: FC<WindowProps> = ({
@@ -40,8 +35,8 @@ const Window: FC<WindowProps> = ({
     onClose,
     children,
 }: WindowProps) => (
-    <Container className={className} onClick={() => onClick()}>
-        <TitleBar name={name} onClose={onClose} />
+    <Container className={className} onClick={onClick}>
+        <WindowBar name={name} onClose={onClose} />
         {children}
     </Container>
 );
